Speed Up `#find_by` Lookup Using Nokogiri
Rather than iterating through the transformed Ruby objects, use Nokogiri
to select the relevant XML node directly and create the transformed
object.

Improves the speed of `#find_by` as we avoid iterating over every
element for a given type.  We have the XML document in memory, so
searching it via Nokogiri is much faster.